tests: ifix tests
